// hairlab/src/app/models/enums/consultation-type.ts
export enum ConsultationType {

  HAIR_CUT = 'HAIR_CUT',

  HAIR_COLOR = 'HAIR_COLOR',

  HAIR_STYLING = 'HAIR_STYLING',

  SCALP_TREATMENT = 'SCALP_TREATMENT',

  HAIR_RESTORATION = 'HAIR_RESTORATION',

  HAIR_EXTENSION = 'HAIR_EXTENSION',

  HAIR_STRAIGHTENING = 'HAIR_STRAIGHTENING',

  HAIR_PERMING = 'HAIR_PERMING',

  HAIR_REPAIR = 'HAIR_REPAIR',

  HAIR_ANALYSIS = 'HAIR_ANALYSIS'
}